requirements.txt
google-cloud-storage>=2.17.0
google-cloud-aiplatform>=1.66.0
google-adk>=0.2.0
google-genai>=0.8.0
pillow>=10.0.0
python-dotenv>=1.0.0
